Enhance signaling flow and SDP negotiation in SFU
- Introduced `send_signaling_offer` and `on_signaling_answer` for dynamic SDP management in Sender.
- Added `on_signaling_offer` to Receiver for responding to SDP offers with answers.
- Updated `DataChannelMsg` handling in Peer to support offer/answer signaling flow.
- Enhanced error handling with new `DataChannel` and `Renegotiation` MediaError types.
- Refactored `SdpMsgData` to use `u64` for `number` field.

+++ src/sfu/media/data_channel.rs
@@ -55,7 +55,7 @@ impl DataChannelMsg {
 
 #[derive(Serialize, Deserialize, Debug)]
 pub struct SdpMsgData {
-    pub number: u32,
+    pub number: u64,
     pub sdp: String,
 }
